refactor (tests): remove hamcrest assertions in engine/junit-shared (#565)
